Module game_logic::scenes

The render code for various scenes

Overview

This will probably become a messy module over time. Stick your rendering code here

Modules

main_menu

This scene encompasses the main menu system

player_interaction

This scene encompasses all of the game where the player can walk around.

test_fox

This “scene” is used only for testing animation and resource loading It should be removed once the game is being worked on

Structs

SceneRenderDelegate

Delegate for handling rendering. This is a struct to allow for stateful data (like sub-screens) to be set up